+++ regimes/gr/gr.go
@@ -1,0 +1,80 @@
+// Package gr provides the tax region definition for Greece.
+package gr
+
+import (
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/currency"
+	"github.com/invopop/gobl/i18n"
+	"github.com/invopop/gobl/l10n"
+	"github.com/invopop/gobl/regimes/common"
+	"github.com/invopop/gobl/tax"
+)
+
+func init() {
+	tax.RegisterRegime(New())
+}
+
+// Custom keys used typically in meta or codes information.
+const (
+	KeyIAPRVatCategory       cbc.Key = "iapr-vat-category"
+	KeyIAPRVatCategoryIsland cbc.Key = "iapr-vat-category-island"
+)
+
+// Official IAPR codes to include in stamps.
+const (
+	StampIAPRQR       cbc.Key = "iapr-qr"
+	StampIAPRMark     cbc.Key = "iapr-mark"
+	StampIAPRHash     cbc.Key = "iapr-hash"
+	StampIAPRUID      cbc.Key = "iapr-uid"
+	StampIAPRProvider cbc.Key = "iapr-provider"
+)
+
+// New provides the tax region definition
+func New() *tax.Regime {
+	return &tax.Regime{
+		Country:  l10n.GR,
+		Currency: currency.EUR,
+		Name: i18n.String{
+			i18n.EN: "Greece",
+			i18n.EL: "Ελλάδα",
+		},
+		TimeZone: "Europe/Athens",
+		Tags:     invoiceTags,
+		Scenarios: []*tax.ScenarioSet{
+			invoiceScenarios,
+		},
+		Corrections: []*tax.CorrectionDefinition{
+			{
+				Schema: bill.ShortSchemaInvoice,
+				Types: []cbc.Key{
+					bill.InvoiceTypeCreditNote,
+				},
+			},
+		},
+		Validator:  Validate,
+		Calculator: Calculate,
+		Categories: taxCategories,
+		Extensions: extensionKeys,
+	}
+}
+
+// Validate checks the document type and determines if it can be validated.
+func Validate(doc interface{}) error {
+	switch obj := doc.(type) {
+	case *bill.Invoice:
+		return validateInvoice(obj)
+	case *tax.Identity:
+		return validateTaxIdentity(obj)
+	}
+	return nil
+}
+
+// Calculate will attempt to clean the object passed to it.
+func Calculate(doc interface{}) error {
+	switch obj := doc.(type) {
+	case *tax.Identity:
+		return common.NormalizeTaxIdentity(obj)
+	}
+	return nil
+}