+++ regimes/gr/scenarios.go
@@ -1,0 +1,40 @@
+package gr
+
+import (
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/i18n"
+	"github.com/invopop/gobl/regimes/common"
+	"github.com/invopop/gobl/tax"
+)
+
+// Document tag keys
+const (
+	TagIslands cbc.Key = "islands"
+)
+
+var invoiceTags = common.InvoiceTagsWith([]*cbc.KeyDefinition{
+	{
+		Key: TagIslands,
+		Name: i18n.String{
+			i18n.EN: "Islands Reduced Rates",
+			i18n.EL: "Νησιά μειωμένοι συντελεστές",
+		},
+	},
+})
+
+var invoiceScenarios = &tax.ScenarioSet{
+	Schema: bill.ShortSchemaInvoice,
+	List: []*tax.Scenario{
+		// ** Special Messages **
+		// Reverse Charges
+		{
+			Tags: []cbc.Key{tax.TagReverseCharge},
+			Note: &cbc.Note{
+				Key:  cbc.NoteKeyLegal,
+				Src:  tax.TagReverseCharge,
+				Text: "Reverse Charge / Αντίστροφη φόρτιση",
+			},
+		},
+	},
+}