+++ regimes/gr/tax_identity_test.go
@@ -1,0 +1,58 @@
+package gr_test
+
+import (
+	"testing"
+
+	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/l10n"
+	"github.com/invopop/gobl/regimes/gr"
+	"github.com/invopop/gobl/tax"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestValidateTaxIdentity(t *testing.T) {
+	tests := []struct {
+		name string
+		code cbc.Code
+		err  string
+	}{
+		{name: "older than 1999", code: "064677095"},
+		{name: "legal entity", code: "925667500"},
+		{name: "natural person", code: "320806520"},
+		{name: "digit repetition", code: "691063615"},
+		{
+			name: "too long",
+			code: "1234567890123",
+			err:  "invalid format",
+		},
+		{
+			name: "too short",
+			code: "123456",
+			err:  "invalid format",
+		},
+		{
+			name: "non-numeric",
+			code: "12345678A",
+			err:  "invalid format",
+		},
+		{
+			name: "bad checksum",
+			code: "925667503",
+			err:  "checksum mismatch",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tID := &tax.Identity{Country: l10n.GR, Code: tt.code}
+			err := gr.Validate(tID)
+			if tt.err == "" {
+				assert.NoError(t, err)
+			} else {
+				if assert.Error(t, err) {
+					assert.Contains(t, err.Error(), tt.err)
+				}
+			}
+		})
+	}
+}