+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [v0.XX.X] - 2024-XX-XX
+
+### Added
+
+- Greece tax regime
+
 ## [v0.79.3] - 2024-06-18
 
 ### Added